+++ pkgs/zebra_datawedge/lib/src/zebra_barcode_scanner.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:combined_barcode_scanner/combined_barcode_scanner.dart';
 import 'package:combined_barcode_scanner_zebra/src/zebra/scan_callback.dart';
 import 'package:combined_barcode_scanner_zebra/src/zebra/zebra_datawedge_controller.dart';
@@ -39,13 +41,13 @@ class ZebraBarcodeScanner implements BarcodeScanner {
     required ScannerConfiguration configuration,
     required ValueChanged<BarcodeScanResult> onScan,
   }) async {
-    _supported = await _controller.isSupported;
+    _supported = await _controller.isControllerSupported;
     if (_supported) {
       await _controller.init(
           profileName, _mapFormats(configuration.enableFormats));
       _controller.scannerCallBack = _ScannerWrapper(onScan);
     }
-    controller = _ZebraController(_controller, enabled: _supported);
+    controller = _ZebraController(_controller, enabled: _supported)..start();
   }
 
   @override
@@ -94,9 +96,12 @@ class _ScannerWrapper implements ScannerCallBack {
   void onError(Exception error) {}
 }
 
-class _ZebraController implements BarcodeScannerController {
+class _ZebraController extends BarcodeScannerController {
   final ZebraDataWedgeController _scanner;
   final bool enabled;
+
+  @override
+  bool get isControllerSupported => enabled;
 
   _ZebraController(
     this._scanner, {